feat: update project to use latest `mc.py` for mc runs - chore: renamed hpc main call to `mc_hpc.py` - chore: remove legacy files - chore: update api call - feat: add `MonteCarloResult` to `config.py` - feat: update `mc.py` to latest successful run scripts - feat: separate methods into appropriate files - chore: update api call for mc results visualization
